requirements.txt
Flask==3.0.0
gunicorn==21.2.0
tensorflow-cpu==2.13.0
numpy==1.24.3